Add files via upload
This program is of IP finder. It will establish the IP address for any URL you enter into the box.
With this you also have a doc file attached which gives the lucid explanation of the code and also lists the terms that you must know before taking up this project.